backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
litellm
google-adk
pydantic>=2,<3
